+++ tests/testthat/setup.R
@@ -1,0 +1,7 @@
+#' Developer sys and svn info
+user_lookup <- dplyr::tribble(
+  ~sys, ~svn,
+  "michaelm", "michaelm",
+  "anderson", "andersone",
+  "graceo", "graceo"
+)